+++ ui/app/src/components/PathNavigator/PathNavigator.tsx
@@ -54,6 +54,8 @@ import { getStoredPathNavigator } from '../../utils/state';
 import { useActiveSite } from '../../hooks/useActiveSite';
 import { useActiveUser } from '../../hooks/useActiveUser';
 import { GetChildrenOptions } from '../../models';
+import TranslationOrText from '../../models/TranslationOrText';
+import usePossibleTranslation from '../../hooks/usePossibleTranslation';
 
 interface Menu {
   path?: string;
@@ -68,7 +70,7 @@ interface Menu {
 
 export interface PathNavigatorProps {
   id: string;
-  label: string;
+  label: TranslationOrText;
   rootPath: string;
   sortStrategy?: GetChildrenOptions['sortStrategy'];
   order?: GetChildrenOptions['order'];
@@ -121,15 +123,15 @@ export interface PathNavigatorStateProps {
 // };
 
 export function PathNavigator(props: PathNavigatorProps) {
+  const title = usePossibleTranslation(props.label) ?? '(No name)';
   // region const { ... } = props;
   const {
-    label = '(No name)',
     icon,
     expandedIcon,
     collapsedIcon,
     container,
     rootPath: path,
-    id = label.replace(/\s/g, ''),
+    id = title.replace(/\s/g, ''),
     limit = 10,
     locale,
     excludes,
@@ -394,7 +396,7 @@ export function PathNavigator(props: PathNavigatorProps) {
         itemsByPath={itemsByPath}
         icon={expandedIcon && collapsedIcon ? (state.collapsed ? collapsedIcon : expandedIcon) : icon}
         container={container}
-        title={label}
+        title={title}
         onChangeCollapsed={onChangeCollapsed}
         onHeaderButtonClick={state.collapsed ? void 0 : onHeaderButtonClick}
         onCurrentParentMenu={onCurrentParentMenu}