+++ ui/app/src/components/PathNavigatorTree/PathNavigatorTree.tsx
@@ -61,6 +61,8 @@ import SystemType from '../../models/SystemType';
 import { PathNavigatorTreeItemProps } from './PathNavigatorTreeItem';
 import { UNDEFINED } from '../../utils/constants';
 import SimpleAjaxError from '../../models/SimpleAjaxError';
+import usePossibleTranslation from '../../hooks/usePossibleTranslation';
+import TranslationOrText from '../../models/TranslationOrText';
 
 export interface PathNavigatorTreeProps
   extends Pick<
@@ -68,7 +70,7 @@ export interface PathNavigatorTreeProps
     'showNavigableAsLinks' | 'showPublishingTarget' | 'showWorkflowState' | 'showItemMenu'
   > {
   id: string;
-  label: string;
+  label: TranslationOrText;
   rootPath: string;
   sortStrategy?: GetChildrenOptions['sortStrategy'];
   order?: GetChildrenOptions['order'];
@@ -129,10 +131,10 @@ interface Menu {
 // };
 
 export function PathNavigatorTree(props: PathNavigatorTreeProps) {
+  const title = usePossibleTranslation(props.label);
   // region const { ... } = props;
   const {
-    label,
-    id = props.label.replace(/\s/g, ''),
+    id = title.replace(/\s/g, ''),
     excludes,
     limit = 10,
     icon,
@@ -334,7 +336,7 @@ export function PathNavigatorTree(props: PathNavigatorTreeProps) {
     <>
       <PathNavigatorTreeUI
         classes={{ header: classes?.header }}
-        title={label}
+        title={title}
         active={active}
         icon={expandedIcon && collapsedIcon ? (state.collapsed ? collapsedIcon : expandedIcon) : icon}
         container={container}